refactor: reorganise components into feature-based folders and update import paths
