+++ frontend/index.html
@@ -5,7 +5,7 @@
     <link
       rel="icon"
       type="image/svg+xml"
-      href="./public/assets/closed-book.svg"
+      href="/assets/closed-book.svg"
     />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>ChapterRewind</title>
