+++ frontend/src/components/Book/BowAnimation.jsx
