+++ frontend/src/components/Book/VideoCarousel.jsx
